Move coolmaster coordinator to its own file
