+++ homeassistant/components/coolmaster/__init__.py
@@ -1,18 +1,13 @@
 """The Coolmaster integration."""
-import logging
-
 from pycoolmasternet_async import CoolMasterNet
 
-from homeassistant.components.climate import SCAN_INTERVAL
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_HOST, CONF_PORT, Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
 from .const import CONF_SWING_SUPPORT, DATA_COORDINATOR, DATA_INFO, DOMAIN
-
-_LOGGER = logging.getLogger(__name__)
+from .coordinator import CoolmasterDataUpdateCoordinator
 
 PLATFORMS = [Platform.CLIMATE, Platform.BINARY_SENSOR, Platform.BUTTON, Platform.SENSOR]
 
@@ -60,25 +55,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     if unload_ok:
         hass.data[DOMAIN].pop(entry.entry_id)
     return unload_ok
-
-
-class CoolmasterDataUpdateCoordinator(DataUpdateCoordinator):
-    """Class to manage fetching Coolmaster data."""
-
-    def __init__(self, hass, coolmaster):
-        """Initialize global Coolmaster data updater."""
-        self._coolmaster = coolmaster
-
-        super().__init__(
-            hass,
-            _LOGGER,
-            name=DOMAIN,
-            update_interval=SCAN_INTERVAL,
-        )
-
-    async def _async_update_data(self):
-        """Fetch data from Coolmaster."""
-        try:
-            return await self._coolmaster.status()
-        except OSError as error:
-            raise UpdateFailed from error
